fix: Return error as sent from the UI instead of throwing a new one (#4610)
## Explanation

When signatures are rejected the UI sends some extra information in the
error data property. This information is stripped off when the error is
re-thrown in the controller. This change aims as throwing the error back
to the UI as received.

Removes the @metamask/rpc-errors package from the Signature controller
as this is not used anymore.
## References

<!--
Are there any issues that this pull request is tied to? Are there other
links that reviewers should consult to understand these changes better?

For example:

* Fixes #12345
* Related to
[#2703](https://github.com/MetaMask/MetaMask-planning/issues/2703)
-->

## Changelog

None


## Checklist

- [ ] I've updated the test suite for new or updated code as appropriate
- [ ] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [ ] I've highlighted breaking changes using the "BREAKING" category
above as appropriate