Add way to view pkg changes since latest release (#1390)
## Description

<!--
Thanks for your contribution! Take a moment to answer these questions so
that reviewers have the information they need to properly understand
your changes:

* What is the current state of things and why does it need to change?
* What is the solution your changes offer and how does it work?
* What packages are you updating?
* Are you introducing a breaking change to a package (renaming or
removing a part of a public interface)?
-->

When creating a new release, it can be helpful to be able to view the
changes that have been made to a package since its most recent release.
This commit adds a package script, `since-latest-release`, which can be
used to accomplish this.

This script is simply a shortcut for a commit range, so you can plug any
Git command that takes a commit range into it. For instance, to view the
list of latest commits for `@metamask/assets-controllers`, you can use
`log`:

yarn workspace @metamask/accounts-controller run since-latest-release
log

Or to view all changes, you can use `diff`:

yarn workspace @metamask/accounts-controller run since-latest-release
diff

Or you can leave an argument off to just get the names of commits:

yarn workspace @metamask/accounts-controller run since-latest-release

## Testing

Try running the command above, using a package that has new changes. Try
to break it.

## Changes

<!--
Pretend that you're updating a changelog. How would you categorize your
changes?

CATEGORY is one of:

- BREAKING
- ADDED
- CHANGED
- DEPRECATED
- REMOVED
- FIXED

(Security-related changes should go through the Security Advisory
process.)
-->

(No user-facing changes; this is a developer-only change.)

## References

<!--
Are there any issues or other links that reviewers should consult to
understand this pull request better? For instance:

* Fixes #12345
* See: #67890
-->

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation for new or updated code as appropriate
(note: this will usually be JSDoc)
- [x] I've highlighted breaking changes using the "BREAKING" category
above as appropriate

---------

Co-authored-by: Salah-Eddine Saakoun <salah-eddine.saakoun@consensys.net>
Co-authored-by: cryptodev-2s <109512101+cryptodev-2s@users.noreply.github.com>