Feat: Block Client Side C2 Requests by Managing a hashed C2 Request Blocklist  (#4526)
## Explanation

This PR implements new logic to manage a Command and Control (C2)
request blocklist within the Phishing Controller. It introduces updates
to both the PhishingController and PhishingDetector to efficiently
handle and validate URLs against a hashed request blocklist.

### Solution
The proposed changes add a requestBlocklist to the Phishing Controller
and Phishing Detector. This includes methods for updating the blocklist,
checking if a request URL's domain is blocked, and incorporating these
checks into the existing phishing detection workflow. This ensures that
any requests to known malicious C2 domains are identified and blocked.

### Notable Changes
Introduction of requestBlocklist to list types.
Implementation of isBlockedRequest in PhishingController.
Update to #updateStalelist to fetch and include the request blocklist.
Addition of isMaliciousRequestDomain in PhishingDetector for URL
validation against the blocklist.

## Changelog

### `@metamask/phishing-controller`

- **ADDED**: `requestBlocklist` type to ListTypes.
- **ADDED**: `isBlockedRequest` method to PhishingController.
- **ADDED**: Logic to update and check `requestBlocklist` in
`#updateStalelist`.
- **ADDED**: `isMaliciousRequestDomain` method to PhishingDetector.
- **ADDED**: Handling of `requestBlocklist` in PhishingDetector
configuration.
- **ADDED**: `sha256` and `toHex` imports from `ethereum-cryptography`.
- **ADDED**: `sha256Hash` function to generate SHA-256 hash of a domain.

---------

Co-authored-by: Jacob Lebowitz <augmentedmode@MacBook-Pro.lan>
Co-authored-by: Prithpal Sooriya <prithpal.sooriya@consensys.net>
Co-authored-by: Harry <409H@users.noreply.github.com>
Co-authored-by: Mark Stacey <markjstacey@gmail.com>
Co-authored-by: OGPoyraz <omergoktugpoyraz@gmail.com>
Co-authored-by: Elliot Winkler <elliot.winkler@gmail.com>
Co-authored-by: Dovydas Stankevicius <dovydas.stankevicius.10@gmail.com>
Co-authored-by: Jongsun Suh <jongsun.suh@icloud.com>
Co-authored-by: Matthew Walsh <matthew.walsh@consensys.net>
Co-authored-by: cryptodev-2s <109512101+cryptodev-2s@users.noreply.github.com>
Co-authored-by: Kanthesha Devaramane <kanthesha.devaramane@consensys.net>
Co-authored-by: Michele Esposito <34438276+mikesposito@users.noreply.github.com>
Co-authored-by: Mathieu Artu <mathieu.artu@consensys.net>