fix: `SelectedNetworkController`  network stateChange listener (#4638)
## Explanation

<!--
Thanks for your contribution! Take a moment to answer these questions so
that reviewers have the information they need to properly understand
your changes:

* What is the current state of things and why does it need to change?
* What is the solution your changes offer and how does it work?
* Are there any changes whose purpose might not obvious to those
unfamiliar with the domain?
* If your primary goal was to update one package but you found you had
to update another one along the way, why did you do so?
* If you had to upgrade a dependency, why did you do so?
-->

Currently, `SelectedNetworkController` subscribes to the
`NetworkController:stateChange` event to make sure that network client
ids associated with managed domains stay valid. Though, we are currently
checking on the specific immer patch applied to the NetworkController
state: this means that based on how the state draft was manipulated on
the origin controller, we may or may not capture the change and react to
it.

To fix it, now `SelectedNetworkController` will look at the actual
content of the updated state, instead of relying on immer patches. To
reduce the amount of event instances received by the controller, a
selector is applied to the subscription, filtering out irrelevant
instances.

## References

<!--
Are there any issues that this pull request is tied to? Are there other
links that reviewers should consult to understand these changes better?

For example:

* Fixes #12345
* Related to #67890
-->

## Changelog

<!--
If you're making any consumer-facing changes, list those changes here as
if you were updating a changelog, using the template below as a guide.

(CATEGORY is one of BREAKING, ADDED, CHANGED, DEPRECATED, REMOVED, or
FIXED. For security-related issues, follow the Security Advisory
process.)

Please take care to name the exact pieces of the API you've added or
changed (e.g. types, interfaces, functions, or methods).

If there are any breaking changes, make sure to offer a solution for
consumers to follow once they upgrade to the changes.

Finally, if you're only making changes to development scripts or tests,
you may replace the template below with "None".
-->

### `@metamask/network-controller`

- **ADDED**: Added `getNetworkConfigurations`,
`getAvailableNetworkClientIds` and `selectAvailableNetworkClientIds`
selectors
  - These new selectors can be applied to messenger event subscriptions

### `@metamask/selected-network-controller`

No consumer-faced changes

## Checklist

- [ ] I've updated the test suite for new or updated code as appropriate
- [ ] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [ ] I've highlighted breaking changes using the "BREAKING" category
above as appropriate