Release 195.0.0 (#4657)
## Explanation

This is a release candidate for v195.0.0.
- `@metamask/profile-sync-controller@0.3.0`
- `@metamask/notification-services-controller@0.3.0`

<!--
Thanks for your contribution! Take a moment to answer these questions so
that reviewers have the information they need to properly understand
your changes:

* What is the current state of things and why does it need to change?
* What is the solution your changes offer and how does it work?
* Are there any changes whose purpose might not obvious to those
unfamiliar with the domain?
* If your primary goal was to update one package but you found you had
to update another one along the way, why did you do so?
* If you had to upgrade a dependency, why did you do so?
-->

## References

<!--
Are there any issues that this pull request is tied to? Are there other
links that reviewers should consult to understand these changes better?

For example:

* Fixes #12345
* Related to #67890
-->

## Changelog

```md
### Changed

- Bump `@metamask/profile-sync-controller` from `^0.2.1` to `^0.3.0` ([#4643](https://github.com/MetaMask/core/pull/4657))
- Bump `@metamask/notification-services-controller` from `^0.2.1` to `^0.3.0` ([#4643](https://github.com/MetaMask/core/pull/4657))
```

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've highlighted breaking changes using the "BREAKING" category
above as appropriate