feat: native scrypt support on Profile Sync Controller (#4656)
## Explanation

This PR implements a pattern of using a React Native scrypt to
encryption while dealing with React native client. This was necessary
due the poor performance for mobile devices relying on the node
implementation. This improves the performance for mobile clientes in a
order of 20x.

## References

<!--
Are there any issues that this pull request is tied to? Are there other
links that reviewers should consult to understand these changes better?

For example:

* Fixes #12345
* Related to #67890
-->

## Changelog

<!--
If you're making any consumer-facing changes, list those changes here as
if you were updating a changelog, using the template below as a guide.

(CATEGORY is one of BREAKING, ADDED, CHANGED, DEPRECATED, REMOVED, or
FIXED. For security-related issues, follow the Security Advisory
process.)

Please take care to name the exact pieces of the API you've added or
changed (e.g. types, interfaces, functions, or methods).

If there are any breaking changes, make sure to offer a solution for
consumers to follow once they upgrade to the changes.

Finally, if you're only making changes to development scripts or tests,
you may replace the template below with "None".
-->

### `@metamask/profile-sync-controller`

- **ADDED**: Optional `NativeScrypt` to `UserStorageController`
- **ADDED**: Switching logic in `Encryption` to use either `NativeScypt`
or default to the original `@noble-hashes/scrypt-async`

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've highlighted breaking changes using the "BREAKING" category
above as appropriate

---------

Co-authored-by: Prithpal Sooriya <prithpal.sooriya@consensys.net>