feat: network-syncing and user storage controller integration (#4687)
## Explanation

This is a follow up on https://github.com/MetaMask/core/pull/4685, and
adds the controller integration for the network mutation syncs.

NOTE - we are currently using mock/temporary events that are not yet
exposed on the network controller. We will add these network events in
an upcoming PR.

## References

https://consensyssoftware.atlassian.net/browse/NOTIFY-1032

## Changelog

<!--
If you're making any consumer-facing changes, list those changes here as
if you were updating a changelog, using the template below as a guide.

(CATEGORY is one of BREAKING, ADDED, CHANGED, DEPRECATED, REMOVED, or
FIXED. For security-related issues, follow the Security Advisory
process.)

Please take care to name the exact pieces of the API you've added or
changed (e.g. types, interfaces, functions, or methods).

If there are any breaking changes, make sure to offer a solution for
consumers to follow once they upgrade to the changes.

Finally, if you're only making changes to development scripts or tests,
you may replace the template below with "None".
-->

### `@metamask/proflile-sync-controller`

- **ADDED**: temporarily added non-existing
`NetworkController:networkAdded`; `NetworkController:networkChanged`;
and `NetworkController:networkDeleted` events.
  - These will be provided in a future PR.
- **ADDED**: add `isNetworkSyncingEnabled` environment switch to
`UserStorageController` to control when we enable network syncing.
- **ADDED**: add `startNetworkSyncing()` to initialise and listen to all
the events required for network syncing.

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've highlighted breaking changes using the "BREAKING" category
above as appropriate