test: remove unused test (#4703)
## Explanation

We have a mock and a test that is not used anymore. It still passed the
CI, but actually doesn't test anything.

## References

<!--
Are there any issues that this pull request is tied to? Are there other
links that reviewers should consult to understand these changes better?

For example:

* Fixes #12345
* Related to #67890
-->

## Changelog

### `@metamask/profile-sync-controller`

- **REMOVED**: unused mock function.

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've highlighted breaking changes using the "BREAKING" category
above as appropriate