chore: Update codeowners to match EM agreed ownership (#4750)
## Explanation
We, the MetaMask team (holistically), have recently spent time
discussing code ownership around the packages in core. As a team we have
aligned on which teams own which packages/controllers within core. This
PR is to reflect that alignment. Additionally, in order to not make
releasing challenging for the Wallet Framework (who are the stewards of
the core repo) we have added this team as codeowners of the manifest
file and the changelog file for all packages.

For a little context being code owners doesn’t mean you are the only
ones who work in that code but rather that you will take responsibility
for making sure that the code is maintained, healthy, up to date and
that others contributing to it do so in a manner that is aligned with
our guidelines, processes and principles.

## References
- [Internal google sheets
doc](https://docs.google.com/spreadsheets/d/1fukzs7OBZBnswPTLysGaZO7I7uBOGYFpfXEK3Krkn5o/edit?usp=sharing)
teams used to align can be found in the google drive
- Additional discussions were in the #mmig-engineering-mgmt channel on
slack

## Checklist

- [ n/a ] I've updated the test suite for new or updated code as
appropriate
- [ n/a ] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [ n/a ] I've highlighted breaking changes using the "BREAKING"
category above as appropriate
- [ n/a ] I've prepared draft pull requests for clients and consumer
packages to resolve any breaking changes