fix: remove includeDuplicateSymbolAssets filter (#4768)
## Explanation

What motivates this change is this user report
https://github.com/MetaMask/metamask-extension/issues/27614

Based on discussions we had with teams; seems like initially, we added
the `includeDuplicateSymbolAssets` because there were concerns around
duplicate symbols being scammy and the increased size of tokenList.

For the duplicate symbols being scammy concern; i think our usage of
`occurrencesFloor=3` is helping us filter out any scam tokens.
Having tokens with the same symbol is not unusual and it does not
necessarily mean that the token is a scam.

For the tokenList size; we thought that; today; the tokenList we get
without `includeDuplicateSymbolAssets` has around 4500 tokens. With the
`includeDuplicateSymbolAssets` it has around 4000. Which does not seem
like it should have a huge impact. We thought that token the tokenList
size would increase regardless in future, and if users add their tokens
to lists and expect to see them we should not be filtering them out even
if they have duplicates.


## References

<!--
Are there any issues that this pull request is tied to?
Are there other links that reviewers should consult to understand these
changes better?
Are there client or consumer pull requests to adopt any breaking
changes?

For example:

* Fixes #12345
* Related to #67890
-->

## Changelog

<!--
If you're making any consumer-facing changes, list those changes here as
if you were updating a changelog, using the template below as a guide.

(CATEGORY is one of BREAKING, ADDED, CHANGED, DEPRECATED, REMOVED, or
FIXED. For security-related issues, follow the Security Advisory
process.)

Please take care to name the exact pieces of the API you've added or
changed (e.g. types, interfaces, functions, or methods).

If there are any breaking changes, make sure to offer a solution for
consumers to follow once they upgrade to the changes.

Finally, if you're only making changes to development scripts or tests,
you may replace the template below with "None".
-->

### `@metamask/assets-controllers`

- **Removed**: Removed param `includeDuplicateSymbolAssets` in our api
call to token-api.


## Checklist

- [ ] I've updated the test suite for new or updated code as appropriate
- [ ] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [ ] I've highlighted breaking changes using the "BREAKING" category
above as appropriate
- [ ] I've prepared draft pull requests for clients and consumer
packages to resolve any breaking changes