feat: :sparkles: add the ability to view draft content for Contentful (#4790)
## Explanation

This PR introduces the ability for the
`fetchFeatureAnnouncementNotifications` function, within the
`notification-services-controller`, to fetch draft content from
Contentful. This is made possible by passing a `previewToken` parameter.
If present, the function will call the correct endpoint and use the
token. This change is particularly useful for teams that need to test
the quality of content in Contentful before its publication.

## References


https://consensyssoftware.atlassian.net/browse/NOTIFY-1159?atlOrigin=eyJpIjoiMTQ3YmE3ZGQwYWY2NGUwOWE2NGMxYzE4NWVlNzk1Y2UiLCJwIjoiaiJ9

## Changelog

### `@metamask/notification-services-controller`

- **ADDED**: The `fetchFeatureAnnouncementNotifications` function now
accepts an optional previewToken parameter

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've highlighted breaking changes using the "BREAKING" category
above as appropriate
- [x] I've prepared draft pull requests for clients and consumer
packages to resolve any breaking changes